@@ -5,13 +5,10 @@
   "author": "CollMot Robotics Ltd",
   "homepage": "",
   "license": "MIT",
-  "main": "lib/index.js",
-  "directories": {
-    "lib": "lib",
-    "test": "__tests__"
-  },
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "files": [
-    "lib"
+    "dist"
   ],
   "publishConfig": {
     "registry": "https://npm.collmot.com"
@@ -21,10 +18,13 @@
     "url": "https://github.com/skybrush-io/skybrush.js"
   },
   "scripts": {
+    "build": "npx tsc",
+    "clean": "rimraf dist/** *.tgz",
+    "prepack": "npm run build",
     "test": "npx jest"
   },
   "peerDependencies": {
-    "electron": "^28.0 || ^29.0 || ^30.0 || ^31.0 || ^32.0 || ^33.0 || ^34.0 || ^35.0 || ^36.0 || ^37.0 || ^38.0 || ^39.0"
+    "electron": ">=39.0"
   },
   "dependencies": {
     "commander": "^11.0.0",
@@ -32,6 +32,6 @@
     "electron-window-state": "^5.0.3"
   },
   "devDependencies": {
-    "electron": "^38.3.0"
+    "electron": "^39"
   }
 }
